refactor repetitive imports
