+++ package.json
@@ -28,6 +28,9 @@
     "examples/react-standalone",
     "examples/html"
   ],
+  "resolutions": {
+    "@web3modal/ui": "packages/core/dist"
+  },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "5.44.0",
     "@typescript-eslint/parser": "5.44.0",
