+++ packages/ui/index.ts
@@ -1,3 +1,7 @@
+import './src/components'
+import './src/partials'
+import './src/views'
+
 export * from './src/components/w3m-modal'
 export * from './src/partials/w3m-account-button'
 export * from './src/partials/w3m-connect-button'
