+++ packages/ui/src/components/index.ts
@@ -1,0 +1,17 @@
+import './w3m-button'
+import './w3m-modal'
+import './w3m-modal-backcard'
+import './w3m-modal-content'
+import './w3m-modal-footer'
+import './w3m-modal-header'
+import './w3m-modal-router'
+import './w3m-modal-toast'
+import './w3m-network-button'
+import './w3m-network-image'
+import './w3m-qrcode'
+import './w3m-search-input'
+import './w3m-spinner'
+import './w3m-text'
+import './w3m-view-all-wallets-button'
+import './w3m-wallet-button'
+import './w3m-wallet-image'
