+++ packages/ui/src/components/w3m-button/index.ts
@@ -2,9 +2,8 @@ import type { TemplateResult } from 'lit'
 import { html, LitElement } from 'lit'
 import { customElement, property } from 'lit/decorators.js'
 import { classMap } from 'lit/directives/class-map.js'
-import { getConditionalValue } from '../../utils/UiHelpers'
 import { global } from '../../utils/Theme'
-import '../w3m-text'
+import { getConditionalValue } from '../../utils/UiHelpers'
 import styles, { dynamicStyles } from './styles'
 
 type Variant = 'fill' | 'ghost'
