+++ packages/ui/src/components/w3m-modal-router/index.ts
@@ -3,21 +3,8 @@ import { RouterCtrl } from '@web3modal/core'
 import { html, LitElement } from 'lit'
 import { customElement, state } from 'lit/decorators.js'
 import { animate } from 'motion'
-import '../../partials/w3m-walletconnect-qr'
 import { global } from '../../utils/Theme'
 import { getShadowRootElement } from '../../utils/UiHelpers'
-import '../../views/w3m-coinbase-extension-connector-view'
-import '../../views/w3m-coinbase-mobile-connector-view'
-import '../../views/w3m-connect-wallet-view'
-import '../../views/w3m-desktop-connector-view'
-import '../../views/w3m-get-wallet-view'
-import '../../views/w3m-help-view'
-import '../../views/w3m-injected-connector-view'
-import '../../views/w3m-metamask-connector-view'
-import '../../views/w3m-qrcode-view'
-import '../../views/w3m-select-network-view'
-import '../../views/w3m-wallet-explorer-view'
-import '../../views/w3m-wallet-filter-view'
 import styles from './styles.css'
 
 @customElement('w3m-modal-router')
