+++ packages/ui/src/components/w3m-text/index.ts
@@ -2,7 +2,6 @@ import { html, LitElement } from 'lit'
 import { customElement, property } from 'lit/decorators.js'
 import { classMap } from 'lit/directives/class-map.js'
 import { global } from '../../utils/Theme'
-import '../w3m-spinner'
 import styles, { dynamicStyles } from './styles'
 
 type Variant =
