+++ packages/ui/src/partials/index.ts
@@ -1,0 +1,6 @@
+import './w3m-account-button'
+import './w3m-connect-button'
+import './w3m-core-button'
+import './w3m-desktop-wallet-selection'
+import './w3m-mobile-wallet-selection'
+import './w3m-walletconnect-qr'
