+++ packages/ui/src/partials/w3m-connect-button/index.ts
@@ -2,8 +2,6 @@ import { ModalCtrl } from '@web3modal/core'
 import { html } from 'lit'
 import { customElement, property, state } from 'lit/decorators.js'
 import { classMap } from 'lit/directives/class-map.js'
-import '../../components/w3m-spinner'
-import '../../components/w3m-text'
 import { WALLET_CONNECT_ICON } from '../../utils/Svgs'
 import { global } from '../../utils/Theme'
 import ThemedElement from '../../utils/ThemedElement'
