+++ packages/ui/src/partials/w3m-desktop-wallet-selection/index.ts
@@ -9,13 +9,6 @@ import {
 } from '@web3modal/core'
 import { html, LitElement } from 'lit'
 import { customElement } from 'lit/decorators.js'
-import '../../components/w3m-modal-content'
-import '../../components/w3m-modal-footer'
-import '../../components/w3m-modal-header'
-import '../../components/w3m-text'
-import '../../components/w3m-view-all-wallets-button'
-import '../../components/w3m-wallet-button'
-import '../../partials/w3m-walletconnect-qr'
 import { COPY_ICON, DESKTOP_ICON, MOBILE_ICON, SCAN_ICON } from '../../utils/Svgs'
 import { global } from '../../utils/Theme'
 import { handleUriCopy } from '../../utils/UiHelpers'
