+++ packages/ui/src/partials/w3m-mobile-wallet-selection/index.ts
@@ -8,10 +8,6 @@ import {
 } from '@web3modal/core'
 import { html, LitElement } from 'lit'
 import { customElement } from 'lit/decorators.js'
-import '../../components/w3m-modal-content'
-import '../../components/w3m-modal-header'
-import '../../components/w3m-view-all-wallets-button'
-import '../../components/w3m-wallet-button'
 import { getOptimisticNamePreset } from '../../utils/Presets'
 import { QRCODE_ICON } from '../../utils/Svgs'
 import { global } from '../../utils/Theme'
