+++ packages/ui/src/views/index.ts
@@ -1,0 +1,12 @@
+import './w3m-coinbase-extension-connector-view'
+import './w3m-coinbase-mobile-connector-view'
+import './w3m-connect-wallet-view'
+import './w3m-desktop-connector-view'
+import './w3m-get-wallet-view'
+import './w3m-help-view'
+import './w3m-injected-connector-view'
+import './w3m-metamask-connector-view'
+import './w3m-qrcode-view'
+import './w3m-select-network-view'
+import './w3m-wallet-explorer-view'
+import './w3m-wallet-filter-view'
