+++ packages/ui/src/views/w3m-coinbase-mobile-connector-view/index.ts
@@ -1,12 +1,6 @@
 import { ClientCtrl, CoreHelpers, ModalCtrl, OptionsCtrl, ToastCtrl } from '@web3modal/core'
 import { html, LitElement } from 'lit'
 import { customElement, state } from 'lit/decorators.js'
-import '../../components/w3m-button'
-import '../../components/w3m-modal-content'
-import '../../components/w3m-modal-footer'
-import '../../components/w3m-modal-header'
-import '../../components/w3m-qrcode'
-import '../../components/w3m-text'
 import { ARROW_DOWN_ICON, QRCODE_ICON } from '../../utils/Svgs'
 import { global } from '../../utils/Theme'
 import { getErrorMessage } from '../../utils/UiHelpers'
