+++ packages/ui/src/views/w3m-get-wallet-view/index.ts
@@ -1,12 +1,6 @@
 import { CoreHelpers, ExplorerCtrl } from '@web3modal/core'
 import { html, LitElement } from 'lit'
 import { customElement } from 'lit/decorators.js'
-import '../../components/w3m-button'
-import '../../components/w3m-modal-content'
-import '../../components/w3m-modal-footer'
-import '../../components/w3m-modal-header'
-import '../../components/w3m-text'
-import '../../components/w3m-wallet-image'
 import { ARROW_RIGHT_ICON, ARROW_UP_RIGHT_ICON } from '../../utils/Svgs'
 import { global } from '../../utils/Theme'
 import { getCustomWallets } from '../../utils/UiHelpers'
