+++ packages/ui/src/views/w3m-qrcode-view/index.ts
@@ -1,11 +1,5 @@
 import { html, LitElement } from 'lit'
 import { customElement } from 'lit/decorators.js'
-import '../../components/w3m-modal-content'
-import '../../components/w3m-modal-footer'
-import '../../components/w3m-modal-header'
-import '../../components/w3m-text'
-import '../../components/w3m-view-all-wallets-button'
-import '../../partials/w3m-walletconnect-qr'
 import { COPY_ICON } from '../../utils/Svgs'
 import { global } from '../../utils/Theme'
 import { handleUriCopy } from '../../utils/UiHelpers'
