+++ packages/ui/src/views/w3m-wallet-filter-view/index.ts
@@ -2,10 +2,6 @@ import type { DesktopConnectorData } from '@web3modal/core'
 import { CoreHelpers, RouterCtrl } from '@web3modal/core'
 import { html, LitElement } from 'lit'
 import { customElement, state } from 'lit/decorators.js'
-import '../../components/w3m-modal-content'
-import '../../components/w3m-modal-header'
-import '../../components/w3m-search-input'
-import '../../components/w3m-wallet-button'
 import { global } from '../../utils/Theme'
 import { getCustomWallets, handleMobileLinking } from '../../utils/UiHelpers'
 import styles, { dynamicStyles } from './styles'
